requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
Deprecated==1.2.18
idna==3.10
pycparser==2.22
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2